+++ crates/ln-dlc-node/src/node/dlc_channel.rs
@@ -266,27 +266,27 @@ impl<S: TenTenOneStorage + 'static, N: LnDlcStorage + Sync + Send + 'static> Nod
         .await?
     }
 
-    pub fn accept_dlc_channel_collaborative_close(&self, channel_id: DlcChannelId) -> Result<()> {
+    pub fn accept_dlc_channel_collaborative_close(&self, channel_id: &DlcChannelId) -> Result<()> {
         let channel_id_hex = hex::encode(channel_id);
 
         tracing::info!(channel_id = %channel_id_hex, "Accepting DLC channel collaborative close offer");
 
         let dlc_manager = self.dlc_manager.clone();
-        dlc_manager.accept_collaborative_close(&channel_id)?;
+        dlc_manager.accept_collaborative_close(channel_id)?;
 
         Ok(())
     }
 
     pub fn accept_dlc_channel_collaborative_settlement(
         &self,
-        channel_id: DlcChannelId,
+        channel_id: &DlcChannelId,
     ) -> Result<()> {
         let channel_id_hex = hex::encode(channel_id);
 
         tracing::info!(channel_id = %channel_id_hex, "Accepting DLC channel collaborative settlement");
 
         let dlc_manager = self.dlc_manager.clone();
-        let (settle_offer, counterparty_pk) = dlc_manager.accept_settle_offer(&channel_id)?;
+        let (settle_offer, counterparty_pk) = dlc_manager.accept_settle_offer(channel_id)?;
 
         self.event_handler.publish(NodeEvent::SendDlcMessage {
             peer: counterparty_pk,